fix: satisfy subagent manager quality gates
